# Use `LocalVector` instead of `List` as arg of `Dictionary::get_key_list`