An attempt to refactor old interface. It is done last month but I thought the changes are too widespread to be merged so I just leave it aside. Anyway, I decide to upload this to spark some discussion.

As a side effect, we can now reserve the key list which reduce calls to malloc.

~~The function name should definitely be changed, maybe `get_keys`? I don't know.~~

Edit: There are many other `xx_list` like API not returning lists, so the name should be ok.