@KoBeWi 2024-12-12
Is it actually faster? Not sure if LocalVector can be used like this.