@Zylann 2024-12-12
I don't see why LocalVector couldn't be used like this. It has indeed potential of being faster by having the opportunity to do a single allocation (instead of many for each List element) and having better locality when iterating. It might even be possible to re-use the same LocalVector if a function uses `get_key_list` many times.

Though I would also be curious about a benchmark showing the difference it makes.