@YYF233333 2024-12-12
Test function performance, for iter performance please refer to [#99936(comment)](https://github.com/godotengine/godot/pull/99936#issuecomment-2520345383).

Build with `scons production=yes tests=yes`

<details>
<summary>Bench Result</summary>

``` bash
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|               78.08 |       12,806,943.63 |    0.0% |      0.01 | `LocalVector Dictionary size: 1`
|              101.51 |        9,850,804.49 |    0.0% |      0.01 | `LocalVector Dictionary size: 3`
|              278.12 |        3,595,549.48 |    0.5% |      0.01 | `LocalVector Dictionary size: 15`
|              929.89 |        1,075,400.65 |    0.2% |      0.01 | `LocalVector Dictionary size: 63`
|            3,606.68 |          277,263.44 |    0.7% |      0.04 | `LocalVector Dictionary size: 255`
|           14,348.54 |           69,693.53 |    0.9% |      0.18 | `LocalVector Dictionary size: 1023`
|           60,682.36 |           16,479.25 |    0.3% |      0.72 | `LocalVector Dictionary size: 4095`

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|              127.48 |        7,844,460.23 |    0.1% |      0.01 | `List Dictionary size: 1`
|              263.94 |        3,788,777.59 |    0.1% |      0.01 | `List Dictionary size: 3`
|            1,091.49 |          916,182.44 |    0.2% |      0.01 | `List Dictionary size: 15`
|            4,362.40 |          229,231.49 |    0.3% |      0.05 | `List Dictionary size: 63`
|           17,739.08 |           56,372.70 |    0.3% |      0.21 | `List Dictionary size: 255`
|           73,570.13 |           13,592.47 |    0.3% |      0.88 | `List Dictionary size: 1023`
|          300,427.36 |            3,328.59 |    0.2% |      3.59 | `List Dictionary size: 4095`
```
</details>

<details>
<summary>Bench Code</summary>

```C++
#define ANKERL_NANOBENCH_IMPLEMENT
#include "nanobench.h"
#include "core/templates/list.h"
#include "core/templates/local_vector.h"
#include "core/variant/dictionary.h"
#include <ostream>
#include <sstream>

class Test
{
    Variant k;
    Dictionary dict;

public:
    void bench_get_key()
    {
        List<Variant> keys;
        dict.get_key_list(&keys);
        for (auto key : keys)
        {
            k = key;
        }
    }

    void insert_to_size(int size)
    {
        for (int i = dict.size(); i < size; i++)
        {
            Variant v = i;
            dict[v] = v;
        }
    }

    static void bench()
    {
        Test test;
        for (int i = 1; i < 4096 + 1; i *= 4)
        {
            int n = i == 1 ? 1 : i-1;
            std::ostringstream name;
            name << "Dictionary size: " << n;
            test.insert_to_size(n);
            auto bench = ankerl::nanobench::Bench()
                .minEpochIterations(1000);
            bench.run("List " + name.str(), [&]{
                test.bench_get_key();
            });
        }
    }
};
```
</details>