@Repiteo 2024-12-12
It's entirely possible I don't understand the technical details, but why is `LocalVector` passed as a reference by argument? Passing by argument made sense when it was a `List` pointer, but wouldn't a return value be more suitable for this new context?