@Zylann 2024-12-12
Unless I'm mistaken, returning LocalVector by value would cause an unnecessary copy of that vector (and therefore another malloc), while passing by reference avoids that copy.