@AThousandShips 2024-12-12
Since c++17 return values like these are guaranteed to perform [copy-elision](https://en.cppreference.com/w/cpp/language/copy_elision) so there should be no copying or moving into the result value, so as long as the value is only used for the return value, and the output is constructed, i.e. `LocalVector<Variant> foo = dict.get_key_list()`, there should be no additional memory operations but the result is constructed in place