@Zylann 2024-12-12
Is that guaranteed even though `get_key_list` is not `inline`?