@AThousandShips 2024-12-12
Yes it's not related to inlining, it's handled by passing the output variable in an argument I believe, so in essence it'd be like passing by reference but automated, quoting the page:
> This sometimes means that even when the language syntax visually suggests a copy/move (e.g. copy initialization), no copy/move is performed — which means the type need not have an accessible copy/move constructor at all.