@Repiteo 2024-12-12
Though now that you mention it, this *could* be inlined if declared in `variant.h`, similar to the `Array` iterator functions