@YYF233333 2024-12-13
> Though now that you mention it, this _could_ be inlined if declared in `variant.h`, similar to the `Array` iterator functions

Needs some tweak, because `DictionaryPrivate` is defined in `.cpp`.