@YYF233333 2024-12-13
> Since c++17 return values like these are guaranteed to perform [copy-elision](https://en.cppreference.com/w/cpp/language/copy_elision) so there should be no copying or moving into the result value

I'm not so convinced after reading some material, the examples they use are too simple, while we are doing non-trival work on the `LocalVector`. Better check the asm to see if it is true.