@YYF233333 2024-12-13
Change the signature. An extra benifit is we can eliminate most of the local variables, leading to a bit more readable code.