@Zylann 2024-12-13
Does profiling give the same results with the change?