@YYF233333 2024-12-13
```bash
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|               78.31 |       12,770,574.23 |    0.0% |      0.01 | `LocalVector Dictionary size: 1`
|              105.07 |        9,517,548.10 |    0.2% |      0.01 | `LocalVector Dictionary size: 3`
|              284.93 |        3,509,644.76 |    0.1% |      0.01 | `LocalVector Dictionary size: 15`
|              950.47 |        1,052,115.03 |    0.5% |      0.01 | `LocalVector Dictionary size: 63`
|            3,609.96 |          277,011.52 |    0.7% |      0.04 | `LocalVector Dictionary size: 255`
|           14,493.03 |           68,998.70 |    0.4% |      0.17 | `LocalVector Dictionary size: 1023`
|           61,365.76 |           16,295.73 |    0.3% |      0.74 | `LocalVector Dictionary size: 4095`
```

A bit slower, but the difference is small enough I think.