@YYF233333 2025-01-10
I'd like to link this PR back to https://github.com/godotengine/godot-proposals/issues/11526, this is an typical example of the first and second point in the proposal.