@YYF233333 2025-01-10
Update a MRP evaluating performance from gdscript side: 
[test-dictionary_get_key_list.zip](https://github.com/user-attachments/files/18380166/test-dictionary_get_key_list.zip)

master:
```bash
801 ms
796 ms
801 ms
799 ms
801 ms
800 ms
795 ms
797 ms
801 ms
798 ms
average: 798 ms
```
this PR:
```bash
662 ms
660 ms
667 ms
668 ms
662 ms
666 ms
666 ms
661 ms
665 ms
666 ms
average: 664 ms
```