@YYF233333 2025-04-07
> Some of these cases should probably just iterate the `Dictionary` without creating a separate `LocalVector`.

Those have been addressed at PR introducing `Dictionary` iterator. Remaining cases in this PR need to duplicate keys (for sorting or modify dict in loop).